+++ lib/slack-ruby-bot/commands/about.rb
@@ -1,10 +1,10 @@
 module SlackRubyBot
   module Commands
     class Default < Base
-      command 'default'
       command 'about'
+      match(/^(?<bot>\w*)$/)
 
-      def self.call(data, _command, _arguments)
+      def self.call(data, _match)
         send_message_with_gif data.channel, SlackRubyBot::ABOUT, 'selfie'
       end
     end