+++ lib/slack-ruby-bot/commands/help.rb
@@ -1,7 +1,7 @@
 module SlackRubyBot
   module Commands
     class Help < Base
-      def self.call(data, _command, _arguments)
+      def self.call(data, _match)
         send_message_with_gif data.channel, 'See https://github.com/dblock/slack-ruby-bot, please.', 'help'
       end
     end