+++ lib/slack-ruby-bot/commands/hi.rb
@@ -1,7 +1,7 @@
 module SlackRubyBot
   module Commands
     class Hi < Base
-      def self.call(data, _command, _arguments)
+      def self.call(data, _match)
         send_message_with_gif data.channel, "Hi <@#{data.user}>!", 'hi'
       end
     end