+++ lib/slack-ruby-bot/commands/unknown.rb
@@ -1,7 +1,9 @@
 module SlackRubyBot
   module Commands
     class Unknown < Base
-      def self.call(data, _command, _arguments)
+      match(/^(?<bot>\w*)[\s]*(?<expression>.*)$/)
+
+      def self.call(data, _match)
         send_message_with_gif data.channel, "Sorry <@#{data.user}>, I don't understand that command!", 'idiot'
       end
     end